fastapi_patient_api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
pydantic
redis
aioredis
python-dotenv